# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /signals:
    get:
      operationId: getSignals
      tags: [signals]
      summary: Get signal history
      description: Returns a list of recently generated signals
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
...
components:
  schemas:
...
    Signal:
      type: object
      properties:
        id:
          type: integer
        market:
          type: string
        symbol:
          type: string
        digit:
          type: integer
        price:
          type: number
        signalType:
          type: string
          enum: [OVER, UNDER, EVEN, ODD, RISE, FALL]
        confidence:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - market
        - symbol
        - digit
        - price
        - signalType
        - confidence
        - createdAt
...